requirements.txt
Weverse==1.0.3
discord.py>=1.6.0
asyncpg==0.23.0
aiohttp==3.7.4.post0
aiofiles==0.7.0
